api_simple/requirements.txt
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
async_fastapi_jwt_auth==0.5.1
asyncpg==0.27.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
cryptography==40.0.2
Django==4.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
Faker==18.3.0
fastapi==0.95.0
greenlet==2.0.2
grpcio==1.51.3
grpcio-tools==1.51.3
h11==0.14.0
httptools==0.5.0
idna==3.4
MarkupSafe==2.1.2
orjson==3.8.8
passlib==1.7.4
protobuf==4.22.1
psycopg2-binary==2.9.6
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.6
Pygments==2.14.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0
redis==4.5.2
requests==2.28.2
rich==12.6.0
rsa==4.9
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.7
sqlparse==0.4.3
starlette==0.26.1
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
Werkzeug==2.2.3
